+++ lib/devise/webauthn/url_helpers.rb
@@ -23,7 +23,10 @@ module Devise
     module UrlHelpers
       {
         passkeys: [nil],
-        passkey: [nil, :new]
+        passkey: [nil, :new],
+        two_factor_authentication: [nil, :new],
+        second_factor_webauthn_credentials: [nil],
+        second_factor_webauthn_credential: [nil, :new]
       }.each do |route, actions|
         %i[path url].each do |path_or_url|
           actions.each do |action|