+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add new `webauthn_two_factor_authenticatable` module for enabling 2FA using WebAuthn credentials.
+
 ## [v0.1.2](https://github.com/cedarcode/devise-webauthn/compare/v0.1.1...v0.1.2/) - 2025-12-03
 
 ### Fixed